services/analysis/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seismosis-analysis"
version = "0.1.0"
requires-python = ">=3.11"
description = "Seismosis analysis service — ML magnitude refinement, ETAS aftershock detection, enrichment"
dependencies = [
    "confluent-kafka>=2.3.0",
    "fastavro>=1.9.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
    "redis[hiredis]>=5.0.0",
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-timeout>=2.3.0",
    "testcontainers[redpanda,postgresql,redis]>=4.3.0",
    "pytest-cov>=5.0.0",
    "requests-mock>=1.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 300
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
